+++ game_patch/misc/misc.h
@@ -11,4 +11,5 @@ bool multi_join_game(const rf::NetAddr& addr, const std::string& password);
 void ui_get_string_size(int* w, int* h, const char* s, int s_len, int font_num);
 void g_solid_render_ui();
 bool tc_mod_is_loaded();
+bool af_rfl_version(int version);
 void evaluate_restrict_disable_ss();