+++ game_patch/object/item.cpp
@@ -1,6 +1,10 @@
 #include <patch_common/FunHook.h>
 #include <patch_common/CodeInjection.h>
+#include <patch_common/AsmWriter.h>
+#include <xlog/xlog.h>
 #include "../rf/item.h"
+#include "../rf/misc.h"
+#include "../rf/multi.h"
 #include "../rf/weapon.h"
 #include "../multi/server.h"
 
@@ -60,8 +64,23 @@ CodeInjection item_create_sort_injection{
     },
 };
 
+CodeInjection game_level_init_pre_patch{
+    0x00435BDC, [](auto& regs) {
+
+        if (!rf::is_multi) {
+            xlog::warn("init powerups");
+            rf::multi_powerup_destroy_all();
+        }
+    }
+};
+
 void item_do_patch()
 {
+    // allow picking up powerups in SP - super health and super armour working. amp/invuln TBD
+    AsmWriter(0x0045AAFD).jmp(0x0045AB11); // allow item_touch_multi_amp in SP
+    AsmWriter(0x0048012B).jmp(0x00480135); // allow multi_powerup_add in SP
+    game_level_init_pre_patch.install();   // initialize powerup vars in SP
+
     // Allow overriding weapon items count value
     item_touch_weapon_hook.install();