+++ game_patch/os/commands.cpp
@@ -85,9 +85,36 @@ ConsoleCommand2 level_info_cmd{
             rf::console::print("Name: {}", rf::level.name);
             rf::console::print("Author: {}", rf::level.author);
             rf::console::print("Date: {}", rf::level.level_date);
-            std::string version_text = (rf::level.version < 200) ? "Official" : "Non-Official";
-            rf::console::print("RFL Version: {} ({})", version_text, rf::level.version);
-        } else {
+            rf::console::print("Has Skybox? {}", rf::level.has_skyroom);
+            rf::console::print("Hardness: {}", rf::level.default_rock_hardness);
+            rf::console::print("Ambient Light: {}, {}, {}",
+                rf::level.ambient_light.red, rf::level.ambient_light.green, rf::level.ambient_light.blue);
+            rf::console::print("Distance Fog: {}, {}, {}, near clip: {}, far clip: {}",
+                rf::level.distance_fog_color.red, rf::level.distance_fog_color.green, rf::level.distance_fog_color.blue,
+                rf::level.distance_fog_near_clip, rf::level.distance_fog_far_clip);
+
+            std::string version_text;
+            if (rf::level.version == 175) {
+                version_text = "Official - PS2 retail";
+            }
+            else if (rf::level.version == 180) {
+                version_text = "Official - PC retail";
+            }
+            else if (rf::level.version == 200) {
+                version_text = "Community - RF/PF/DF";
+            }
+            else if (rf::level.version > 0 && rf::level.version < 200) {
+                version_text = "Official - internal";
+            }            
+            else if (rf::level.version >= 300) {
+                version_text = "Community - AF";
+            }
+            else {
+                version_text = "Unsupported";
+            }
+            rf::console::print("RFL File Version: {} ({})", rf::level.version, version_text);
+        }
+        else {
             rf::console::print("No level loaded!");
         }
     },